tests/toolbox.inc.sh: allow test to run in container
keyctl seccomp is not namespaced in a Docker container, see:
https://docs.docker.com/engine/security/seccomp/#significant-syscalls-blocked-by-the-default-profile

The keyctl command works properly, it just means that the keyctl command inside
the container will manipulate the host machine's /proc/keys, but it's not
allowed to read or manipulate it directly inside the container (it will be
always empty inside). That's why the grep command will always exit without any
sleep, but this will not allow enought time for the kernel to destroy the key,
which means the assert "keyctl rdescribe" will fail because it runs when the key
is still there.

We can't bind mount /proc or /proc/keys into the container, so there is no way
to read /proc/keys from the inside (we could mount it to another path, but we
would still need to fix the test in that case). So the best we can do is just
wait until the key should definitely be destroyed.